+++ src/os/bluestore/PMEMDevice.h
@@ -20,8 +20,8 @@
 #include <atomic>
 
 #include "os/fs/FS.h"
-#include "os/fs/aio.h"
 #include "include/interval_set.h"
+#include "aio.h"
 #include "BlockDevice.h"
 
 class PMEMDevice : public BlockDevice {
